Make custom equality comparer
